requirements.txt
langchain
langchain-core
langchain-community
langchain-openai

python-dotenv
ipykernel
pydantic
numexpr

faiss-cpu
unstructured
langchain-pinecone
pinecone-client
streamlit

langchain-ollama
langdetect
deep-translator

gradio